Move Schedulable interface to the tasks package
