+++ src/main/java/duke/tasks/Schedulable.java
@@ -1,4 +1,4 @@
-package duke.utils;
+package duke.tasks;
 
 import java.time.LocalDateTime;
