+++ src/main/java/duke/tasks/TaskManager.java
@@ -4,7 +4,6 @@ import com.google.gson.reflect.TypeToken;
 import com.google.gson.typeadapters.RuntimeTypeAdapterFactory;
 import duke.utils.PersistentList;
 import duke.utils.ResourceHandler;
-import duke.utils.Schedulable;
 
 import java.text.MessageFormat;
 import java.util.ArrayList;
